constants: rename to courseConstants, moved service constants
- rename utils/constants -> utils/course-constants
- email service -> services/email/constants
- stripe service -> moved inline with stripe service methods
- refactored usages
